# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad Assessment API",
    "version": "0.2.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Create Assessment",
        "operationId": "create_assessment_api_v1_assessments_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_assessment_api_v1_assessments_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "List Assessments",
        "description": "Assessment history, most recent first.",
        "operationId": "list_assessments_api_v1_assessments_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assessment"
                  },
                  "title": "Response List Assessments Api V1 Assessments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assessments/{assessment_id}": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Get Assessment",
        "operationId": "get_assessment_api_v1_assessments__assessment_id__get",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brands": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "List Brands",
        "operationId": "list_brands_api_v1_brands_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Brand"
                  },
                  "type": "array",
                  "title": "Response List Brands Api V1 Brands Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "brands"
        ],
        "summary": "Create Brand",
        "operationId": "create_brand_api_v1_brands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Brand"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brands/{brand_id}": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "Get Brand",
        "operationId": "get_brand_api_v1_brands__brand_id__get",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Brand"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "brands"
        ],
        "summary": "Delete Brand",
        "operationId": "delete_brand_api_v1_brands__brand_id__delete",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brands/{brand_id}/documents": {
      "get": {
        "tags": [
          "brands"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_brands__brand_id__documents_get",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDocument"
                  },
                  "title": "Response List Documents Api V1 Brands  Brand Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "brands"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_brands__brand_id__documents_post",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_brands__brand_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comparisons": {
      "post": {
        "tags": [
          "comparisons"
        ],
        "summary": "Create Comparison",
        "description": "Score several variants of the same ad and rank them.",
        "operationId": "create_comparison_api_v1_comparisons_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_comparison_api_v1_comparisons_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdMetadata": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "ad_type": {
            "type": "string",
            "title": "Ad Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "platform",
          "industry",
          "ad_type"
        ],
        "title": "AdMetadata",
        "description": "What the user tells us about where the ad will run."
      },
      "AdScorecard": {
        "properties": {
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Criterion"
            },
            "type": "object",
            "title": "Scores",
            "description": "One 0-10 score per criterion. All six are required."
          },
          "overall_score": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Overall Score"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback",
            "description": "A short paragraph of overall assessment."
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "At most five concrete, actionable changes."
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations",
            "description": "Filenames of the guidance passages the judgement rests on."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scores",
          "overall_score",
          "feedback"
        ],
        "title": "AdScorecard",
        "description": "The scoring contract.\n\nDoubles as the provider output schema, so a provider physically cannot\nreturn a shape this application does not understand."
      },
      "Assessment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "metadata": {
            "$ref": "#/components/schemas/AdMetadata"
          },
          "scorecard": {
            "$ref": "#/components/schemas/AdScorecard"
          },
          "visual_analysis": {
            "$ref": "#/components/schemas/VisualAnalysis"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/RetrievedChunk"
            },
            "type": "array",
            "title": "Context"
          },
          "image_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Sha256"
          },
          "provider_info": {
            "type": "object",
            "title": "Provider Info",
            "description": "Which models and prompt version produced this, for cache keying."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "metadata",
          "scorecard",
          "visual_analysis"
        ],
        "title": "Assessment",
        "description": "A complete assessment: what was asked, what was seen, what was judged."
      },
      "Body_create_assessment_api_v1_assessments_post": {
        "properties": {
          "ad_image": {
            "type": "string",
            "format": "binary",
            "title": "Ad Image"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "ad_type": {
            "type": "string",
            "title": "Ad Type"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          }
        },
        "type": "object",
        "required": [
          "ad_image",
          "platform",
          "industry",
          "ad_type"
        ],
        "title": "Body_create_assessment_api_v1_assessments_post"
      },
      "Body_create_comparison_api_v1_comparisons_post": {
        "properties": {
          "ad_images": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Ad Images"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "ad_type": {
            "type": "string",
            "title": "Ad Type"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          }
        },
        "type": "object",
        "required": [
          "ad_images",
          "platform",
          "industry",
          "ad_type"
        ],
        "title": "Body_create_comparison_api_v1_comparisons_post"
      },
      "Body_upload_document_api_v1_brands__brand_id__documents_post": {
        "properties": {
          "document": {
            "type": "string",
            "format": "binary",
            "title": "Document"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "Body_upload_document_api_v1_brands__brand_id__documents_post"
      },
      "Brand": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "Brand",
        "description": "An advertiser whose own guidelines augment the global best practices."
      },
      "Comparison": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "metadata": {
            "$ref": "#/components/schemas/AdMetadata"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ComparisonEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "metadata",
          "entries"
        ],
        "title": "Comparison",
        "description": "Several variants of the same ad, judged together and ranked."
      },
      "ComparisonEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank",
            "description": "1 is best."
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "The uploaded filename, for identification."
          },
          "assessment": {
            "$ref": "#/components/schemas/Assessment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rank",
          "label",
          "assessment"
        ],
        "title": "ComparisonEntry",
        "description": "One variant's placement in a comparison."
      },
      "CreateBrandRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateBrandRequest"
      },
      "Criterion": {
        "type": "string",
        "enum": [
          "attention",
          "clarity",
          "targeting",
          "cta",
          "branding",
          "value"
        ],
        "title": "Criterion"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "knowledge_ready": {
            "type": "boolean",
            "title": "Knowledge Ready"
          },
          "vision_provider": {
            "type": "string",
            "title": "Vision Provider"
          },
          "scoring_provider": {
            "type": "string",
            "title": "Scoring Provider"
          }
        },
        "type": "object",
        "required": [
          "status",
          "knowledge_ready",
          "vision_provider",
          "scoring_provider"
        ],
        "title": "HealthResponse"
      },
      "KnowledgeDocument": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "brand_id": {
            "type": "string",
            "title": "Brand Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_sha256": {
            "type": "string",
            "title": "Content Sha256"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "brand_id",
          "filename",
          "content_sha256",
          "chunk_count",
          "created_at"
        ],
        "title": "KnowledgeDocument",
        "description": "One ingested brand document."
      },
      "RetrievedChunk": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunk": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "source"
        ],
        "title": "RetrievedChunk",
        "description": "One passage of marketing guidance pulled from the knowledge store."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisualAnalysis": {
        "properties": {
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors",
            "description": "Dominant colors, hex where legible."
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Key copy detected in the creative.",
            "default": ""
          },
          "layout": {
            "type": "string",
            "title": "Layout",
            "description": "Composition and focal areas.",
            "default": ""
          },
          "cta": {
            "type": "string",
            "title": "Cta",
            "description": "Visible call-to-action text, empty if none.",
            "default": ""
          },
          "insights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Insights",
            "description": "Short strength/risk observations."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "VisualAnalysis",
        "description": "Structured description of the creative, produced by the vision model."
      }
    }
  }
}